+++ src/users/users.module.ts
@@ -4,11 +4,13 @@ import { XmlApiModule } from 'src/xml-api/xml-api.module';
 import { UsersController } from './controllers/users.controller';
 import { UsersService } from './services/users.service';
 import { EncodingModule } from 'src/encoding/encoding.module';
+import { UsernamesService } from './services/usernames.service';
+import { UsernamesController } from './controllers/usernames.controller';
 
 @Module({
   imports: [HttpModule, XmlApiModule, EncodingModule],
   exports: [UsersService],
-  providers: [UsersService],
-  controllers: [UsersController],
+  providers: [UsersService, UsernamesService],
+  controllers: [UsersController, UsernamesController],
 })
 export class UsersModule {}