+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 
 - Author profiles now include information about whether the author's account has been locked.
+- Implemented `GET /usernames` endpoint. This endpoint returns filterable lists of usernames.
 
 ## [2.1.0] - 2023-10-20